+++ src/main/java/com/stereowalker/survive/Survive.java
@@ -8,6 +8,7 @@ import com.stereowalker.survive.compat.OriginsCompat;
 import com.stereowalker.survive.config.Config;
 import com.stereowalker.survive.config.ServerConfig;
 import com.stereowalker.survive.config.TemperatureConfig;
+import com.stereowalker.survive.config.WellbeingConfig;
 import com.stereowalker.survive.core.cauldron.SCauldronInteraction;
 import com.stereowalker.survive.events.SurviveEvents;
 import com.stereowalker.survive.json.ArmorJsonHolder;
@@ -70,6 +71,7 @@ public class Survive extends MinecraftMod {
 	public static final String MOD_ID = "survive";
 	public static final Config CONFIG = new Config();
 	public static final TemperatureConfig TEMPERATURE_CONFIG = new TemperatureConfig();
+	public static final WellbeingConfig WELLBEING_CONFIG = new WellbeingConfig();
 	public static boolean isPrimalWinterLoaded;
 	public static final ItemConsummableDataManager consummableReloader = new ItemConsummableDataManager();
 	public static final PotionDrinkDataManager potionReloader = new PotionDrinkDataManager();
@@ -93,6 +95,7 @@ public class Survive extends MinecraftMod {
 		ConfigBuilder.registerConfig(ServerConfig.class);
 		ConfigBuilder.registerConfig(CONFIG);
 		ConfigBuilder.registerConfig(TEMPERATURE_CONFIG);
+		ConfigBuilder.registerConfig(WELLBEING_CONFIG);
 		final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
 		modEventBus.addListener(this::setup);
 		modEventBus.addListener(this::clientRegistries);