+++ src/main/java/com/stereowalker/survive/config/Config.java
@@ -134,12 +134,6 @@ public class Config implements ConfigObject {
 	@UnionConfig.Comment(comment = {"Disabling this will disable this mods nutrition system","The nutrition system might be pretty buggy, so proceed with caution"})
 	public boolean nutrition_enabled = false;
 	
-	//
-	//Wellbeing
-	@UnionConfig.Entry(group = "Well Being" , name = "Enable Well Being (Beta)", type = Type.COMMON)
-	@UnionConfig.Comment(comment = {"Disabling this will disable this mods well being system","The well being system might be pretty buggy, so proceed with caution"})
-	public boolean wellbeing_enabled = false;
-	
 	
 	//
 	//Misc