+++ changelog.md
@@ -1,3 +1,7 @@
+[4.1.0]
+- Gave temperature its dedicated config folder
+- tweaked the behavior of temperature to fit in with 1.17. "Use Experimental Temperature System" has to be enabled to access this change
+
 [4.0.0]
 - Updated to 1.17.1
 - Fixed energized effect doing nothing